Module Section

This is the module comment. Eventually, sections won't be allowed in it.

Empty section

Text only

Foo bar.

Aside only

Value only

let foo: unit;

Empty section

within a comment

and one with a nested section

This section title has markup

But links are impossible thanks to the parser, so we never have trouble rendering a section title in a table of contents – no link will be nested inside another link.